Prefix attachment columns from CSV export with a local attachment file path.

attachment_link(data_tbl, att_path = "media", att_contains = "photo")

Arguments

data_tbl
The downloaded submissions from `ruODK::submission_export()` read into a `tibble` by `readr::read_csv()`.

att_path
A local path, default: "media" (as per .csv.zip export). Selected columns of the dataframe (containing attchment filenames) are prefixed with `att_path`, thus turning them into relative paths.

att_contains
A shared part of attachment fieldnames, default: "photo". Columns of the dataframe are selected by the presence of `att_contains`.

Value

The dataframe with attachment columns modified to contain relative paths to the downloaded attachment files.

See also

Examples

if (FALSE) {
t <- tempdir()
# Set default credentials, see vignette "setup"
ruODK::ru_setup(
  svc = paste0(
    "https://sandbox.central.opendatakit.org/v1/projects/14/",
    "forms/build_Flora-Quadrat-0-2_1558575936.svc"
  ),
  un = "me@email.com",
  pw = "..."
)

# Predict filenames (with knowledge of form)
fid <- get_default_fid()
fid_csv <- fs::path(t, glue::glue("{fid}.csv"))
fid_csv_tae <- fs::path(t, glue::glue("{fid}-taxon_encounter.csv"))

# Download the zip file
se <- ruODK::submission_export(
  local_dir = t,
  overwrite = FALSE,
  verbose = TRUE
)

# Unpack the zip file
f <- unzip(se, exdir = t)
fs::dir_ls(t)

# Prepend attachments with media/ to turn into relative file paths
data_quadrat <- fid_csv %>%
  readr::read_csv(na = c("", "NA", "na")) %>%
  janitor::clean_names(.) %>%
  attachment_link(.) %>%
  parse_datetime(tz = "Australia/Perth")
}