List all submissions of one form.

attachment_list(iid, pid = get_default_pid(), fid = get_default_fid(),
  url = get_default_url(), un = get_default_un(),
  pw = get_default_pw())

Arguments

iid
The `instance_id`, a UUID, as returned by `submission_list`.

pid
The numeric ID of the project, e.g.: 2. Default: get_default_pid(). Set default pid through ru_setup(pid="..."). See vignette("Setup", package = "ruODK").

fid
The alphanumeric form ID, e.g. "build_Spotlighting-0-8_1559885147". Default: get_default_fid(). Set default fid through ru_setup(fid="..."). See vignette("Setup", package = "ruODK").

url
The ODK Central base URL without trailing slash.#' Default: get_default_url(). Set default url through ru_setup(url="..."). See vignette("Setup", package = "ruODK").

un
The ODK Central username (an email address). Default: get_default_un(). Set default un through ru_setup(un="..."). See vignette("Setup", package = "ruODK").

pw
The ODK Central password. Default: get_default_pw(). Set default pw through ru_setup(pw="..."). See vignette("Setup", package = "ruODK").

Value

A tibble containing some high-level details of the submission attachments. One row per submission attachment, columns are submission attributes:

* name: The attachment filename, e.g. 12345.jpg * exists: Whether the attachment for that submission exists on the server.

Details

When a Submission is created, either over the OpenRosa or the REST interface, its XML data is analysed to determine which file attachments it references: these may be photos or video taken as part of the survey, or an audit/timing log, among other things. Each reference is an expected attachment, and these expectations are recorded permanently alongside the Submission. With this subresource, you can list the expected attachments, see whether the server actually has a copy or not, and download, upload, re-upload, or clear binary data for any particular attachment.

You can retrieve the list of expected Submission attachments at this route, along with a boolean flag indicating whether the server actually has a copy of the expected file or not. If the server has a file, you can then append its filename to the request URL to download only that file (see below).

See also

Examples

if (FALSE) {
# Set default credentials, see vignette "setup"
ruODK::ru_setup(
  svc = paste0(
    "https://sandbox.central.opendatakit.org/v1/projects/14/",
    "forms/build_Flora-Quadrat-0-2_1558575936.svc"
  ),
  un = "me@email.com",
  pw = "..."
)

sl <- submission_list()

al <- attachment_list(sl$instance_id[[1]])
al %>% knitr::kable(.)

# attachment_list returns a tibble
class(al)
# > c("tbl_df", "tbl", "data.frame")

# Submission attributes are the tibble's columns
names(al)
# > "name" "exists"
}