Parse a form_schema into a tibble of fields with name, type, and path.

form_schema_parse(fs, path = "Submissions", verbose = FALSE)

Arguments

fs
The output of form_schema as nested list

path
The base path for form fields. Default: "Submissions". `form_schema_parse()` recursively steps into deeper nesting levels, which are reflected as separate OData tables. The returned value in `path` reflects the XForm group name, which translates to separate screens in ODK Collect. Non-repeating form groups will be flattened out into the main Submissions table. Repeating groups are available as separate OData tables.

verbose
Whether to display debug messages or not (default)

Details

The `form_schema` is a nested list of lists containing the form definition. The form definition consists of fields (with a type and name), and form groups, which are rendered as separate ODK Collect screens. Form groups in turn can also contain form fields.

`form_schema_parse` recursively unpacks the form and extracts the name and type of each field. This information then can be used to inform the user which columns require `parse_datetime`, `attachment_get`, or `attachment_link`, respectively.

See also

Examples

if (FALSE) {
# Option 1: in two steps
fs <- form_schema(flatten = FALSE) # Default, but shown for clarity
fsp <- form_schema_parse(fs)

# Option 2: in one go
fsp <- form_schema(parse = TRUE)

fsp

# Attachments: use `attachment_get` on each of
fsp %>% dplyr::filter(type == "binary")

# dateYime: use `parse_datetime` on each of
fsp %>% dplyr::filter(type == "dateTime")

# Point location: will be split into lat/lon/alt/acc
fsp %>% dplyr::filter(type == "geopoint")
}