JSON form schema for an ODK Central form.

Usage

fq_form_schema

Format

The output of ruODK::form_schema(), a tibble with columns "type", "name" and "path" and one row per form field.

Source

See system.file("extdata", "FloraQuadrat04.xml", package = "ruODK") and ruODK::form_schema().

Details

The parsed form schema of an ODK Central form.

This data is kept up to date with the data used in vignettes and package tests. The data is comprised of test records with nonsensical data. The forms used to capture this data are development versions of real-world forms.

This data is used to build vignettes offline and without the need for credentials to an ODK Central server. The test suite ensures that the "canned" data is identical to the "live" data.

See also