Parse all columns containing a certain word in its name to localised datetime.

parse_datetime(df, col_contains = "time", orders = c("YmdHMS",
  "YmdHMSz"), tz = "Australia/Perth")

Arguments

df
A dataframe or tibble.

col_contains
A character string indicating a date/time column. This can be a part of the column name or the whole column name. Default: "time" will match all column names containing "time".

orders
(vector of character) Orders of datetime elements for lubridate. Default: c("YmdHMS", "YmdHMSz").

tz
(character) The timezone string for lubridate. Default: "UTC".

Value

The dataframe with matching columns mutated to lubridate datetimes.

Details

This function wraps a `dplyr::mutate_at()` operation on all datetime columns containing a tell-tale `col_contains` string like "time" (default), "datetime" or "date". The operator using this function will have to know the column names (or commonalities like `"..._time"`).

See also

Examples

if (FALSE) {
tsi <- "data/build_Turtle-Sighting-0-1_1559790020.csv" %>%
  readr::read_csv(na = c("", "NA", "na")) %>%
  janitor::clean_names() %>%
  link_attachments() %>%
  parse_datetime(dt_contains = "Date", tz = "Australia/Perth") %>%
  parse_datetime(tz = "Australia/Perth")
}