Get submissions for a list of submission instance IDs.

Uses get_one_submission on a list of submission instance IDs (iid) as returned from submission_list$instance_id. By giving the list of iid to download explicitly, that list can be modified using information not accessible to ruODK, e.g. iid can be restricted to "only not already downloaded submissions".

Usage

submission_get(
  iid,
  pid = get_default_pid(),
  fid = get_default_fid(),
  url = get_default_url(),
  un = get_default_un(),
  pw = get_default_pw(),
  retries = get_retries()
)

Arguments

iid

A list of submission instance IDs, e.g. from submission_list$instance_id.

pid

The numeric ID of the project, e.g.: 2.

Default: get_default_pid.

Set default pid through ru_setup(pid="...").

See vignette("Setup", package = "ruODK").

fid

The alphanumeric form ID, e.g. "build_Spotlighting-0-8_1559885147".

Default: get_default_fid.

Set default fid through ru_setup(fid="...").

See vignette("Setup", package = "ruODK").

url

The ODK Central base URL without trailing slash.

Default: get_default_url.

Set default url through ru_setup(url="...").

See vignette("Setup", package = "ruODK").

un

The ODK Central username (an email address). Default: get_default_un. Set default un through ru_setup(un="..."). See vignette("Setup", package = "ruODK").

pw

The ODK Central password. Default: get_default_pw. Set default pw through ru_setup(pw="..."). See vignette("Setup", package = "ruODK").

retries

The number of attempts to retrieve a web resource.

This parameter is given to RETRY(times = retries).

Default: 3.

Value

A nested list of submission data.

Details

Forms with submission audit enabled will also receive the submission audit as audit.csv. This will overwrite all previous audit.csv files. To get the combined submission audit logs as one single, concatenated audit.csv file, use submission_export. Note that the submission audit is different from the server audit. The submission audit logs the actions of the enumerator in ODK Collect, whereas the form audit logs the actions of the project manager in ODK Central.

See also

Examples

if (FALSE) {
# Step 1: Setup ruODK with OData Service URL (has url, pid, fid)
ruODK::ru_setup(svc = "...")

# Step 2: List all submissions of form
sl <- submission_list()

# Step 3: Get submissions
subs <- submission_get(sl$instance_id)
}