Get or set ruODK settings.

ru_settings()

get_default_pid()

get_default_fid()

get_default_url()

get_default_un()

get_default_pw()

get_test_url()

get_test_un()

get_test_pw()

get_test_pid()

get_test_fid()

get_test_fid_zip()

get_test_fid_att()

Value

ru_settings prints your default ODK Central project ID, form ID, url, username, and password, and corresponding optional test server settings. ru_setup sets your production and test settings, while get_(default/test)_* get each of those respective settings.

See also

Examples

ru_settings()

#> <ruODK settings>
#>   Default ODK Central Project ID:  1 
#>   Default ODK Central Form ID:  build_Predator-or-Disturbance-1-1_1559789410 
#>   Default ODK Central URL:  https://odkcentral.dbca.wa.gov.au 
#>   Default ODK Central Username:  florian.mayer@dbca.wa.gov.au 
#>   Default ODK Central Password: run ruODK::get_default_pw() to show 
#>   Test ODK Central Project ID: 14 
#>   Test ODK Central Form ID: build_Flora-Quadrat-0-4_1564384341 
#>   Test ODK Central Form ID (ZIP tests): build_Spotlighting-0-6_1558333698 
#>   Test ODK Central Form ID (Attachment tests): build_Flora-Quadrat-0-1_1558330379 
#>   Test ODK Central URL: https://sandbox.central.opendatakit.org 
#>   Test ODK Central Username: Florian.Mayer@dbca.wa.gov.au 
#>   Test ODK Central Password: run ruODK::get_test_pw() to show